Introduce middleware for serial communcations
This commit further splits the console apart from the emulator. The new
"machine" code acts as a middleware between the emulated computer and
the console. The emulated computer should now be easier to port as it
should not depend on any library calls.

The middleware is further more split in two halves, a server and a
client which can communicate through a serial device. This enables a
future port to a micro controler where the emulated cpu runs on one
machine and the console on another.